Change volume text only on start and on haptic pulse
This seems cleaner than updating unnecessarily on every frame; I don't
actually see a clear difference in the profiler. The canvas may well avoid
re-rendering text if it hasn't changed. Or maybe it does it every frame
but it's cheap. Dunno!

The editor UI looks as if an event can be associated
non-programatically, but I couldn't get that to work.